Improving on game logic
updated index.html to include data-sq values for each game square 0-8
data is transferred to the event object on click and stored in
game-logic.js as currentClick.

Working on PATCH request to update the api after each move. the
currentClick will be used to send the data for which square to be
updated.

Storing a variable currentMove in game-logic to keep track of which
player turn it is (x or o)