Making progress on game logic
Created a function checkWinner in game-logic that checks for all 8
instances of a win. If all 9 boxes are selected without a winner then
the result is a tie.

Organized some functions. ran into and fixed a problem with circular
logic where files were requiring and exporting to the same file.