Fixed bug where user could continue to click on board after a win.
setTimeout for mousedown was executing after turning the event listener
off, causing the gameboard to be active after a win. deleted setTimeout
and turned event listener on if move is invalid, and after checking for
a winner. If the move resulted in a winner, event listener will
switch off until new game initiated.